requirements.txt
wget
asyncio
# pyrogram==1.4.16
pyrofork
tgcrypto
oauth2client
psycopg2-binary
sqlalchemy==1.3.20
httplib2
pySmartDL
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
yt-dlp
ffmpeg
tenacity
python-dotenv
